backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
python-dotenv>=1.0.0
boring-semantic-layer>=0.3.7
ibis-framework[bigquery]
pyarrow
db-dtypes
google-auth
google-cloud-bigquery
cachetools>=5.3.0
duckdb==1.1.3
google-cloud-storage>=2.14.0
structlog>=24.1.0
openpyxl>=3.1.0
pytest>=8.0
pytest-asyncio>=0.23
